+++ jquantlib-helpers/src/main/java/org/jquantlib/helpers/FDDividendOptionHelper.java
@@ -162,7 +162,7 @@ public abstract class FDDividendOptionHelper<T extends FDEngineAdapter> extends 
         this.stochProcess = new BlackScholesMertonProcess(new Handle<Quote>(spot), qTS, rTS, volTS);
 
         // obtain a pricing engine and assign to this option :: 3 intervals a day
-        final int timeSteps = exercise.lastDate().sub(referenceDate) * 3;
+        final int timeSteps = (int) (exercise.lastDate().sub(referenceDate) * 3);
         final PricingEngine engine;
         try {
             final Constructor<T> baseConstructor = engineClass.getConstructor(
