+++ jquantlib/src/main/java/org/jquantlib/cashflow/Coupon.java
@@ -120,7 +120,7 @@ public abstract class Coupon extends CashFlow {
                 refPeriodEnd);
     }
 
-    public int accrualDays() {
+    public long accrualDays() {
         return dayCounter().dayCount(accrualStartDate,
                 accrualEndDate);
     }
